Extract async token mocks out so unit tests do not run
